+++ packages/pangea-sdk/pangea/asyncio/services/vault.py
@@ -31,6 +31,9 @@ from pangea.services.vault.models.common import (
     EncryptStructuredResult,
     EncryptTransformRequest,
     EncryptTransformResult,
+    ExportEncryptionAlgorithm,
+    ExportRequest,
+    ExportResult,
     FolderCreateRequest,
     FolderCreateResult,
     GetRequest,
@@ -556,6 +559,7 @@ class VaultAsync(ServiceBaseAsync):
         rotation_frequency: Optional[str] = None,
         rotation_state: Optional[ItemVersionState] = None,
         expiration: Optional[datetime.datetime] = None,
+        exportable: Optional[bool] = None,
     ) -> PangeaResponse[SymmetricGenerateResult]:
         """
         Symmetric generate
@@ -577,6 +581,7 @@ class VaultAsync(ServiceBaseAsync):
                 - `deactivated`
                 - `destroyed`
             expiration (str, optional): Expiration timestamp
+            exportable (bool, optional): Whether the key is exportable or not
 
         Raises:
             PangeaAPIException: If an API Error happens
@@ -616,6 +621,7 @@ class VaultAsync(ServiceBaseAsync):
             rotation_frequency=rotation_frequency,
             rotation_state=rotation_state,
             expiration=expiration,
+            exportable=exportable,
         )
         return await self.request.post(
             "v1/key/generate", SymmetricGenerateResult, data=input.model_dump(exclude_none=True)
@@ -632,6 +638,7 @@ class VaultAsync(ServiceBaseAsync):
         rotation_frequency: Optional[str] = None,
         rotation_state: Optional[ItemVersionState] = None,
         expiration: Optional[datetime.datetime] = None,
+        exportable: Optional[bool] = None,
     ) -> PangeaResponse[AsymmetricGenerateResult]:
         """
         Asymmetric generate
@@ -653,6 +660,7 @@ class VaultAsync(ServiceBaseAsync):
                 - `deactivated`
                 - `destroyed`
             expiration (str, optional): Expiration timestamp
+            exportable (bool, optional): Whether the key is exportable or not
 
         Raises:
             PangeaAPIException: If an API Error happens
@@ -692,6 +700,7 @@ class VaultAsync(ServiceBaseAsync):
             rotation_frequency=rotation_frequency,
             rotation_state=rotation_state,
             expiration=expiration,
+            exportable=exportable,
         )
         return await self.request.post(
             "v1/key/generate", AsymmetricGenerateResult, data=input.model_dump(exclude_none=True)
@@ -711,6 +720,7 @@ class VaultAsync(ServiceBaseAsync):
         rotation_frequency: Optional[str] = None,
         rotation_state: Optional[ItemVersionState] = None,
         expiration: Optional[datetime.datetime] = None,
+        exportable: Optional[bool] = None,
     ) -> PangeaResponse[AsymmetricStoreResult]:
         """
         Asymmetric store
@@ -734,6 +744,7 @@ class VaultAsync(ServiceBaseAsync):
                 - `deactivated`
                 - `destroyed`
             expiration (str, optional): Expiration timestamp
+            exportable (bool, optional): Whether the key is exportable or not
 
         Raises:
             PangeaAPIException: If an API Error happens
@@ -777,6 +788,7 @@ class VaultAsync(ServiceBaseAsync):
             rotation_frequency=rotation_frequency,
             rotation_state=rotation_state,
             expiration=expiration,
+            exportable=exportable,
         )
         return await self.request.post("v1/key/store", AsymmetricStoreResult, data=input.model_dump(exclude_none=True))
 
@@ -792,6 +804,7 @@ class VaultAsync(ServiceBaseAsync):
         rotation_frequency: Optional[str] = None,
         rotation_state: Optional[ItemVersionState] = None,
         expiration: Optional[datetime.datetime] = None,
+        exportable: Optional[bool] = None,
     ) -> PangeaResponse[SymmetricStoreResult]:
         """
         Symmetric store
@@ -814,6 +827,7 @@ class VaultAsync(ServiceBaseAsync):
                 - `deactivated`
                 - `destroyed`
             expiration (str, optional): Expiration timestamp
+            exportable (bool, optional): Whether the key is exportable or not
 
         Raises:
             PangeaAPIException: If an API Error happens
@@ -855,6 +869,7 @@ class VaultAsync(ServiceBaseAsync):
             rotation_frequency=rotation_frequency,
             rotation_state=rotation_state,
             expiration=expiration,
+            exportable=exportable,
         )
         return await self.request.post("v1/key/store", SymmetricStoreResult, data=input.model_dump(exclude_none=True))
 
@@ -1391,5 +1406,51 @@ class VaultAsync(ServiceBaseAsync):
         return await self.request.post(
             "v1/key/decrypt/transform",
             DecryptTransformResult,
+            data=input.model_dump(exclude_none=True),
+        )
+
+    async def export(
+        self,
+        id: str,
+        version: int | None = None,
+        encryption_key: str | None = None,
+        encryption_algorithm: ExportEncryptionAlgorithm | None = None,
+    ) -> PangeaResponse[ExportResult]:
+        """
+        Export
+
+        Export a symmetric or asymmetric key.
+
+        OperationId: vault_post_v1_export
+
+        Args:
+            id: The ID of the item.
+            version: The item version.
+            encryption_key: Public key in pem format used to encrypt exported key(s).
+            encryption_algorithm: The algorithm of the public key.
+
+        Raises:
+            PangeaAPIException: If an API error happens.
+
+        Returns:
+            A `PangeaResponse` where the exported key is returned in the
+            `response.result` field. Available response fields can be found in
+            our [API documentation](https://pangea.cloud/docs/api/vault#export).
+
+        Examples:
+            exp_encrypted_resp = await self.vault.export(
+                id=id,
+                version=1,
+                encryption_key=rsa_pub_key_pem,
+                encryption_algorithm=ExportEncryptionAlgorithm.RSA4096_OAEP_SHA512,
+            )
+        """
+
+        input: ExportRequest = ExportRequest(
+            id=id, version=version, encryption_algorithm=encryption_algorithm, encryption_key=encryption_key
+        )
+        return await self.request.post(
+            "v1/export",
+            ExportResult,
             data=input.model_dump(exclude_none=True),
         )